[Modify] entity 수정 - 코드 주석 수정 - 위치 수정
